refactor(web): emit each backend into its own top-level output directory
Previously the WebGPU build emitted benchmark.js/wasm into the shared
wasm/wasm-mt dir and then relocated it, so the two backends still collided
during the build. Now each backend builds directly into demo/webgl or
demo/webgpu (with wasm/wasm-mt sub-dirs), eliminating the copy/overwrite step.
