+++ web/demo/index-st.html
@@ -13,6 +13,6 @@
 <div id="container" class="container">
     <canvas class="canvas" id="benchmark"></canvas>
 </div>
-<script type="module" src="./wasm/index-st.js"></script>
+<script type="module" src="./webgl/wasm/index-st.js"></script>
 </body>
 </html>
