+++ web/demo/index-st.ts
@@ -17,7 +17,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 import {TGFXBind} from '../lib/tgfx';
-import Benchmark from './wasm/benchmark';
+import Benchmark from './webgl/wasm/benchmark';
 import {ShareData, updateSize, onresizeEvent, startDraw, setCanvasDefaultSize, setupCoordinateConversion} from "./common";
 
 let shareData: ShareData = new ShareData();
@@ -26,7 +26,7 @@ if (typeof window !== 'undefined') {
     window.onload = async () => {
         try {
             setupCoordinateConversion('benchmark');
-            shareData.BenchmarkModule = await Benchmark({ locateFile: (file: string) => './wasm/' + file });
+            shareData.BenchmarkModule = await Benchmark({ locateFile: (file: string) => './webgl/wasm/' + file });
             TGFXBind(shareData.BenchmarkModule);
             let tgfxView = shareData.BenchmarkModule.TGFXView.MakeFrom('#benchmark');
             shareData.tgfxBaseView = tgfxView;
