+++ web/demo/index-webgpu-st.html
@@ -13,6 +13,6 @@
 <div id="container" class="container">
     <canvas class="canvas" id="benchmark"></canvas>
 </div>
-<script type="module" src="./wasm-webgpu/index-webgpu-st.js"></script>
+<script type="module" src="./webgpu/wasm/index-webgpu-st.js"></script>
 </body>
 </html>
