+++ web/demo/index-webgpu.html
@@ -13,6 +13,6 @@
 <div id="container" class="container">
     <canvas class="canvas" id="benchmark"></canvas>
 </div>
-<script type="module" src="./wasm-mt-webgpu/index-webgpu.js"></script>
+<script type="module" src="./webgpu/wasm-mt/index-webgpu.js"></script>
 </body>
 </html>
