+++ web/demo/index.html
@@ -13,6 +13,6 @@
 <div id="container" class="container">
     <canvas class="canvas" id="benchmark"></canvas>
 </div>
-<script type="module" src="./wasm-mt/index.js"></script>
+<script type="module" src="./webgl/wasm-mt/index.js"></script>
 </body>
 </html>
