+++ web/package.json
@@ -8,7 +8,7 @@
   "typings": "types/tgfx.d.ts",
   "scripts": {
     "clean:tgfx": "rimraf lib/ types/",
-    "clean": "npm run clean:tgfx && rimraf --glob demo/build demo/wasm demo/wasm-mt demo/wasm-webgpu demo/wasm-mt-webgpu demo/.*.md5 demo/*.js demo/*.map demo/cmake-build-*",
+    "clean": "npm run clean:tgfx && rimraf --glob demo/webgl demo/webgpu demo/*.js demo/*.map",
     "build:tgfx": "npm run clean:tgfx && tsc -p ./tsconfig.type.json && rollup -c ./script/rollup.tgfx.js",
     "build": "node script/cmake.demo.js -a wasm-mt && npm run build:tgfx && rollup -c ./script/rollup.demo.js --environment ARCH:wasm-mt,BACKEND:webgl",
     "build:debug": "node script/cmake.demo.js -a wasm-mt --debug && npm run build:tgfx && rollup -c ./script/rollup.demo.js --environment ARCH:wasm-mt,BACKEND:webgl",
